+++ www/src/components/diagram.js
@@ -4,7 +4,6 @@ import { Fragment } from "react"
 import { keyframes } from "@emotion/core"
 import { Link, StaticQuery, graphql } from "gatsby"
 
-import { mediaQueries } from "../gatsby-plugin-theme-ui"
 import logo from "../assets/monogram.svg"
 import { GraphQLIcon, ReactJSIcon } from "../assets/tech-logos"
 import FuturaParagraph from "../components/futura-paragraph"
@@ -108,17 +107,11 @@ const boxPadding = { py: 3, px: 4 }
 const SourceItem = ({ children }) => (
   <div
     sx={{
-      boxSizing: `border-box`,
       py: 4,
       px: 5,
       display: `flex`,
-      [mediaQueries.xs]: {
-        flex: `1 1 50%`,
-      },
-      [mediaQueries.sm]: {
-        flex: `1 1 33%`,
-        maxWidth: `33%`,
-      },
+      flex: [null, `1 1 50%`, null, `1 1 33%`],
+      maxWidth: [null, null, null, `33%`],
     }}
   >
     <div
@@ -146,10 +139,10 @@ const ItemTitle = ({ children }) => (
   </h3>
 )
 
-const ItemDescription = ({ children }) => (
+const ItemDescription = ({ children, color }) => (
   <small
     sx={{
-      color: `textMuted`,
+      color: color ? color : `textMuted`,
       display: `block`,
       fontFamily: `system`,
       fontSize: 1,
@@ -181,19 +174,17 @@ const Gatsby = () => (
       src={logo}
       sx={{
         display: `inline-block`,
-        height: t => t.space[8],
+        height: [t => t.space[8], null, null, null, t => t.space[9]],
         margin: 0,
         verticalAlign: `middle`,
         width: `auto`,
-        [mediaQueries.lg]: {
-          height: t => t.space[9],
-        },
       }}
       alt="Gatsby"
     />
     <ItemDescription>
       <small
         sx={{
+          color: `grey.50`,
           display: `block`,
           mt: 2,
           mb: 1,
@@ -236,10 +227,6 @@ const Diagram = () => (
           sx={{
             fontWeight: `heading`,
             mb: 6,
-            mt: 0,
-            [mediaQueries.md]: {
-              mt: 6,
-            },
           }}
         >
           How Gatsby works
@@ -304,7 +291,7 @@ const Diagram = () => (
                 width: `auto`,
               }}
             >
-              <ItemDescription>
+              <ItemDescription color="grey.50">
                 HTML &middot; CSS &middot;
                 {` `}
                 <TechWithIcon icon={ReactJSIcon} height="1.1em">
@@ -325,7 +312,7 @@ const Diagram = () => (
               pb: 5,
             }}
           >
-            <ItemTitle>Static Web Host</ItemTitle>
+            <ItemTitle>Web Hosting</ItemTitle>
             <ItemDescription>
               {staticHosts.map(({ node: staticHost }, index) => (
                 <Fragment key={staticHost.url}>