+++ www/src/components/showcase-details.js
@@ -23,7 +23,7 @@ const gutterDesktop = 8
 
 const styles = {
   link: {
-    color: `link.default`,
+    color: `link.color`,
     textDecoration: `none`,
   },
   prevNextLink: {
@@ -32,7 +32,6 @@ const styles = {
     position: `absolute`,
     top: 280,
     width: 300,
-    transform: `translateX(-75px) rotate(90deg)`,
   },
   prevNextLinkSiteTitle: {
     color: `gatsby`,
@@ -130,9 +129,11 @@ const SourceLink = ({ ...props }) => (
       display: `flex`,
       alignItems: `center`,
       mr: 3,
+      color: `link.color`,
+      width: `100%`,
     }}
   >
-    <GithubIcon sx={{ fontSize: 3, mr: 2 }} />
+    <GithubIcon sx={{ fontSize: 3, mr: 2, color: `link.color` }} />
     Source
   </a>
 )
@@ -194,9 +195,11 @@ const ShowcaseDetails = ({ parent, data, isModal, categories }) => (
               }}
               css={{
                 display: `block`,
-                position: `fixed`,
-                top: `150px`,
-                transform: `translateX(750px)`,
+                order: 2,
+                position: `relative`,
+                [mediaQueries.md]: {
+                  top: `110px`,
+                },
               }}
             >
               <div css={{ margin: `25px` }}>
@@ -215,7 +218,12 @@ const ShowcaseDetails = ({ parent, data, isModal, categories }) => (
                   alt=""
                 />
               </div>
-              <div sx={styles.prevNextLink}>
+              <div
+                sx={{
+                  ...styles.prevNextLink,
+                  transform: `translateX(-75px) rotate(90deg)`,
+                }}
+              >
                 <MdArrowUpward />
                 <div> Next Site in Showcase </div>
                 <div sx={styles.prevNextLinkSiteTitle}>{nextSite.title}</div>
@@ -231,9 +239,12 @@ const ShowcaseDetails = ({ parent, data, isModal, categories }) => (
               }}
               css={{
                 display: `block`,
-                position: `fixed`,
-                top: `150px`,
-                transform: `translateX(-100%)`,
+                order: 1,
+                position: `relative`,
+                [mediaQueries.md]: {
+                  order: 0,
+                  top: `110px`,
+                },
               }}
             >
               <div css={{ margin: `25px` }}>
@@ -270,16 +281,17 @@ const ShowcaseDetails = ({ parent, data, isModal, categories }) => (
         >
           <div
             sx={{
-              alignItems: `center`,
               display: `flex`,
               flexDirection: `column`,
               maxWidth: isModal ? false : 1080,
               margin: isModal ? false : `0 auto`,
+              width: `100%`,
+              order: 1,
             }}
           >
             <div css={{ width: `100%` }}>
-              <Helmet>
-                <title>{data.sitesYaml.title}: Showcase | GatsbyJS</title>
+              <Helmet titleTemplate="%s | GatsbyJS">
+                <title>{`${data.sitesYaml.title}: Showcase`}</title>
                 <meta
                   property="og:image"
                   content={`https://www.gatsbyjs.org${screenshotFile.resize.src}`}
@@ -415,6 +427,7 @@ const ShowcaseDetails = ({ parent, data, isModal, categories }) => (
                         <Link
                           to={`/showcase?${qs.stringify({ filters: [c] })}`}
                           state={{ isModal: true }}
+                          sx={styles.link}
                         >
                           {c}
                         </Link>