+++ www/src/gatsby-plugin-theme-ui/index.js
@@ -24,6 +24,11 @@ for (let b in bp) {
   breakpointsTokens.push(bp[b])
 }
 
+// remove the first breakpoint, `xxs: 0`
+// this made sense for styled-system and using an object
+// to define breakpoints, but not here
+breakpointsTokens.splice(0, 1)
+
 let fontsTokens = {}
 for (let fontFamily in f) {
   fontsTokens[fontFamily] = f[fontFamily].join(`, `)
@@ -42,7 +47,7 @@ const lineHeightsTokens = {
   heading: lh.dense,
 }
 
-const darkBackground = `#111014` // meh
+const darkBackground = `#131217` // meh
 const darkBorder = c.grey[90]
 // const darkBackground = c.purple[90]
 // const darkBorder = c.purple[80]
@@ -100,11 +105,11 @@ const col = {
     comment: `#527713`,
     tag: `#137886`,
     regex: `#dc0437`,
+    remove: c.red[70],
+    add: c.green[80],
     // unchanged
     border: `#faede5`,
     text: `#866c5b`,
-    remove: `#e45c5c`,
-    add: `#4a9c59`,
     punctuation: `#53450e`,
     cssString: `#a2466c`,
     invisibles: `#e0d7d1`,
@@ -206,9 +211,9 @@ const col = {
   modes: {
     dark: {
       background: darkBackground,
-      text: c.grey[20],
-      heading: c.white,
-      textMuted: c.grey[50],
+      text: c.grey[30],
+      heading: c.whiteFade[80],
+      textMuted: c.grey[40],
       banner: hex2rgba(c.purple[90], 0.975),
       muted: c.grey[90],
       icon: {
@@ -251,7 +256,7 @@ const col = {
         keyword: c.magenta[30],
         punctuation: c.whiteFade[70],
         regex: `#d88489`,
-        remove: c.red[60],
+        remove: c.red[40],
         selector: c.orange[30],
         tag: c.teal[60],
         text: c.grey[30],
@@ -365,7 +370,9 @@ export const zIndices = z
 const config = {
   // this enables the color modes feature
   // and is used as the name for the top-level colors object
-  initialColorMode: `light`,
+  initialColorModeName: `light`,
+  // `prefers-color-scheme: dark` media query
+  useColorSchemeMediaQuery: true,
   // borders: borders,
   breakpoints: breakpointsTokens,
   colors: col,
@@ -391,6 +398,22 @@ const config = {
       fontSize: 2,
       py: 2,
       px: 3,
+    },
+  },
+  links: {
+    muted: {
+      fontSize: 1,
+      lineHeight: `solid`,
+      py: 3,
+      "&&": {
+        border: 0,
+        color: `textMuted`,
+        display: `flex`,
+        fontWeight: `body`,
+      },
+      "&&:hover": {
+        color: `link.hoverColor`,
+      },
     },
   },
 }